finaluzation
