+++ src/examples/calc-app/operations.module.ts
@@ -1,0 +1,33 @@
+import { declareModule, injectable } from '../../index';
+import { RootModule, REGISTER_COMMAND_TOKEN } from './calc-root.module';
+
+export const OperationsModule = declareModule({
+  name: 'OperationsModule',
+  providers: [
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['current', (cur) => cur],
+    }),
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['clear', () => 0],
+    }),
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['plus', (cur, inputVals) => inputVals.reduce((acc, val) => acc + val, cur)],
+    }),
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['minus', (cur, inputVals) => inputVals.reduce((acc, val) => acc - val, cur)],
+    }),
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['multiply', (cur, inputVals) => inputVals.reduce((acc, val) => acc * val, cur)],
+    }),
+    injectable({
+      provide: REGISTER_COMMAND_TOKEN,
+      useValue: ['divide', (cur, inputVals) => inputVals.reduce((acc, val) => acc / val, cur)],
+    }),
+  ],
+  imports: [RootModule],
+});
