+++ src/index.ts
@@ -3,3 +3,4 @@ export { injectable } from './module/provider-declaration';
 export { declareModule } from './module/module-declaration';
 export { createToken, modifyToken } from './base/token';
 export type { Token, TokenProvide, TokenDecProvide, TokensDeclarationProvide } from './base/token.types';
+export type { ContainerConfiguration } from './module/container-declaration';
