+++ src/module/container-declaration.ts
@@ -10,10 +10,7 @@ import type { Container } from '../base/container.types';
 import type { Token, TokenDec } from '../base/token.types';
 import type { ModuleDeclaration } from './module.types';
 
-type Configuration = {
-  /**
-   * @todo
-   */
+export type ContainerConfiguration = {
   modules?: ModuleDeclaration[];
   providers: InjectableDeclaration<TodoAny>[];
   parent?: Container;
@@ -45,7 +42,7 @@ const orderModuleProviders = (modules?: ModuleDeclaration[]): InjectableDeclarat
   const uniqProviders = new Set<InjectableDeclaration>();
   const traverseModules = (importedModules?: ModuleDeclaration[]) => {
     importedModules?.forEach((moduleDec) => {
-      const { providers, imports = [], name } = moduleDec;
+      const { providers, imports = [] } = moduleDec;
       traverseModules(imports);
       providers.forEach((providerDec) => {
         uniqProviders.add(providerDec);
@@ -59,21 +56,19 @@ const orderModuleProviders = (modules?: ModuleDeclaration[]): InjectableDeclarat
 };
 
 /**
- * Token for creating scoped providers, each dependency for this provider with different scope will behave different:
- * - 'singleton' dependencies are singleton for parent container scope
- * - 'scoped' dependencies are singleton for each child container
- * - 'transient' dependencies are never the same
+ * Token for creating child containers, an injectable should be passed as the root of this container
  */
 export const CHILD_DI_FACTORY_TOKEN =
   createToken<<A>(childProvider: InjectableDeclaration<A>) => () => A>('inverter:child-di-factory');
 
-export const declareContainer = ({ providers, modules, parent }: Configuration) => {
+export const declareContainer = ({ providers, modules, parent }: ContainerConfiguration) => {
   const container = createBaseContainer(parent);
   const resolvingTokens = new Set<Token<TodoAny>>(); // for cycle dep check
+  const allProviders = [...orderModuleProviders(modules), ...providers];
   const traverseProviders = (provider: InjectableDeclaration<TodoAny>, stack: Token<TodoAny>[]) => {
     provider.inject?.forEach((dec) => {
       const { token, optional } = extractDeclaration(dec);
-      const provided = providers.find((x) => x.provide === token);
+      const provided = allProviders.find((x) => x.provide === token);
       stack.push(token);
 
       if (typeof provided === 'undefined') {
@@ -98,8 +93,6 @@ export const declareContainer = ({ providers, modules, parent }: Configuration) 
         providers: [childProvider],
       }).get(childProvider.provide);
   });
-
-  const allProviders = [...orderModuleProviders(modules), ...providers];
 
   allProviders.forEach((injectableDep) => {
     const { inject, provide, useFactory, useValue, scope } = injectableDep;
